Explicitly specify dependency output file name.
Using "gcc (GCC) 4.1.2 20070925 (Red Hat 4.1.2-33)" in Fedora 8, the
MMD option does not preserve the directory portion of the dependency
even though the -o option is present (i.e., it uses the equivalent of
"$(notdir $*).d" instead of "$*.d" even though the documentation
states that it will preserve the directory portion of the dependency
when the -o option is present).  To get around this behaviour,
explicitly name the dependency file using the -MF option.